Fix ObjectCacheExportStats test failure and remove l1CacheSize config
Summary:
This test timed out because it directly set the cache size as 1073741824 (1024 * 1024 * 1024) bytes. As the entriesLimit = 10, it need to allocate as many as 16711670 placeholders. The test timed out after 10 mins when allocating 7594570th placeholder.

Based on the entriesLimit, 8388608 bytes cache size is enough. The test passed after removing the `l1CacheSize` config.

Reviewed By: therealgymmy

Differential Revision: D36523030

fbshipit-source-id: 36ac8e63db6bd5b66d05d19626bd0cbb61d6c680